backend/pyproject.toml
[project]
name = "finviz-pro-backend"
version = "1.0.0"
description = "Finance Analytics Dashboard Backend API"
authors = [
    { name = "Raymond Oluoch" }
]
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.1",
    "flask-migrate>=4.0.5",
    "flask-sqlalchemy>=3.1.1",
    "flask-restful>=0.3.10",
    "flask-jwt-extended>=4.5.3",
    "flask-limiter>=3.5.0",
    "flask-caching>=2.1.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "pandas>=2.1.3",
    "numpy>=1.26.2",
    "marshmallow>=3.20.1",
    "marshmallow-sqlalchemy>=0.29.0",
    "bcrypt>=4.0.1",
    "passlib>=1.7.4",
    "email-validator>=2.1.0",
    "redis>=5.0.1",
    "gunicorn>=21.2.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "werkzeug>=3.0.0",
    "black>=26.3.1",
    "isort>=8.0.1",
    "autoflake>=2.3.3",
    "autopep8>=2.3.2",
    "phonenumbers>=9.0.25",
    "psutil>=7.2.2",
    "scikit-learn>=1.8.0",
    "flask-restx>=1.3.2",
    "openpyxl>=3.1.5",
    "pytest-mock>=3.15.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-flask>=1.2.0",
    "pytest-xdist>=3.3.1",
    "factory-boy>=3.3.0",
    "faker>=19.13.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "factory-boy>=3.3.3",
    "faker>=40.8.0",
    "finviz-pro-backend[dev]",
    "flake8>=7.3.0",
    "isort>=8.0.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-flask>=1.3.0",
    "pytest-xdist>=3.8.0",
]

[tool.pre-commit]
# Pre-commit configuration
install = true

[tool.pre-commit.hooks]
trailing-whitespace = true
end-of-file-fixer = true
check-yaml = true
check-json = true
check-merge-conflict = true
debug-statements = true
detect-private-key = true

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --tb=short"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "api: API tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
]

